requirements.txt
absl-py==0.8.1
appdirs==1.4.3
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==1.5.0
cachetools==3.1.1
certifi==2020.4.5.1
cffi==1.12.3
chardet==3.0.4
Click==7.0
cupy-cuda90==6.0.0
cycler==0.10.0
Cython==0.29.7
decorator==4.4.0
fastrlock==0.4
ffmpeg==1.4
Flask==1.0.3
funcsigs==1.0.2
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphviz==0.13
grpcio==1.25.0
h5py==2.10.0
html5lib==0.9999999
imageio==2.5.0
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonpatch==1.24
jsonpointer==2.0
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
llvmlite==0.29.0
lmdb==0.94
Mako==1.0.13
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.4
mkl-fft==1.0.12
mkl-random==1.0.2
more-itertools==7.0.0
networkx==2.3
nltk==3.4
numba==0.44.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.1.0
paddlepaddle-gpu==1.5.2.post97
pandas==0.24.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.11.0
prettytable==0.7.2
progress==1.5
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyblur==0.2.3
pycocotools==2.0
pycparser==2.19
pycuda==2019.1
Pygments==2.4.0
pyparsing==2.4.0
pytest==4.0.0
python-dateutil==2.8.0
pytools==2019.1.1
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.0
rarfile==3.1
recordio==0.1.7
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.2.1
Shapely==1.6.4.post2
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
tb-nightly==1.14.0a20190301
tensorboard==1.8.0
tensorboardX==1.7
tensorflow-estimator==2.0.1
tensorflow-gpu==1.8.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
torchfile==0.1.0
torchnet==0.0.4
tornado==6.0.2
tqdm==4.36.1
traitlets==4.3.2
visdom==0.1.8.9
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.4
wrapt==1.11.1
